# Conflicts:
#	backend/src/main/java/pl/sda/poznan/bank/backend/controller/api/v1/BankAccountController.java
#	backend/src/main/java/pl/sda/poznan/bank/backend/repository/BankAccountRepository.java
#	backend/src/main/java/pl/sda/poznan/bank/backend/service/BankAccountService.java
#	backend/src/main/java/pl/sda/poznan/bank/backend/service/impl/BankAccountServiceImpl.java

+++ backend/src/main/java/pl/sda/poznan/bank/backend/model/BankAccount.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.List;
 
 @Entity
 @Getter
@@ -22,12 +23,12 @@ public class BankAccount {
 
     private double balance;
 
-    private  int accountNumber;
+    private  long accountNumber;
 
     public BankAccount() {
     }
 
-    public BankAccount(AccountType accountType, User user, int accountNumber) {
+    public BankAccount(AccountType accountType, User user, long accountNumber) {
         this.accountType = accountType;
         this.user = user;
         this.accountNumber = accountNumber;
